refactor: split postgrest-filters into a directory module
Break the single postgrest-filters.ts into one file per public util
(subset-params-to-search, load-subset-options-to-search, key-columns-to-search,
query-ir-to-search) over a shared common.ts, with index.ts exposing only those
four functions. Keep the module's internal helpers file-local, drop a stray
plugin cache artifact, and gitignore .impeccable/.
