+++ src/postgrest-filters/query-ir-to-search.ts
@@ -1,9 +1,3 @@
-import {
-  extractSimpleComparisons,
-  type LoadSubsetOptions,
-  parseOrderByExpression,
-  type SimpleComparison,
-} from "@tanstack/db"
 import type {
   SerializedExpression as Expression,
   SerializedFrom,
@@ -11,329 +5,14 @@ import type {
   SerializedQueryIR,
   SerializedSelect,
   SerializedWhere,
-} from "./serialize"
-
-type Comparison = { column: string; operator: string; value: string }
-type PostgrestParam =
-  | ({ kind: "column" } & Comparison)
-  | { kind: "group"; filter: string }
-
-interface FilterOptions {
-  mergeIn?: boolean
-  strict?: boolean
-  stripAlias?: boolean
-}
-
-// Only lists and logical groups parse quoted values. Top-level scalar values
-// must remain raw: col=eq."x" would match the quotes themselves.
-const NEEDS_QUOTES = /^$|^\s|\s$|[,()"\\]/
-const quoteValue = (value: unknown): string => {
-  const raw = `${value}`
-  return NEEDS_QUOTES.test(raw)
-    ? `"${raw.replace(/\\/g, "\\\\").replace(/"/g, '\\"')}"`
-    : raw
-}
-
-const SCALAR_OPERATORS = ["eq", "gt", "gte", "lt", "lte", "like", "ilike"]
-
-function renderComparison(
-  expr: Expression,
-  quoteScalars: boolean,
-  options: FilterOptions
-): Comparison | null {
-  if (expr.type !== "func") return null
-  const [left, right] = expr.args
-  if (expr.name === "not" && left) {
-    if (left.type === "func" && left.name === "not" && left.args[0]) {
-      return renderComparison(left.args[0], quoteScalars, options)
-    }
-    const inner = renderComparison(left, quoteScalars, options)
-    if (!inner) return null
-    // TanStack's NOT IN [] excludes nulls; PostgREST's NOT IN () includes them.
-    if (inner.operator === "in" && inner.value === "()") {
-      return { ...inner, operator: "not.is", value: "null" }
-    }
-    return {
-      ...inner,
-      operator: `not.${inner.operator}`,
-    }
-  }
-  if (left?.type !== "ref" || left.path[0] === "$selected") return null
-  const column = (options.stripAlias ? left.path.slice(1) : left.path).join(".")
-  if (!column) return null
-  if (expr.name === "isNull") return { column, operator: "is", value: "null" }
-  if (right?.type !== "val") return null
-
-  if (expr.name === "in") {
-    if (!Array.isArray(right.value)) return null
-    // TanStack's membership test ignores null list members for non-null rows.
-    const values = Array.from(
-      new Set(right.value.filter((value) => value != null))
-    )
-    // PostgREST treats IN ("") as an empty list, even when quoted.
-    if (values.length === 1 && values[0] === "") {
-      return { column, operator: "eq", value: quoteScalars ? '""' : "" }
-    }
-    return {
-      column,
-      operator: "in",
-      value: `(${values.map(quoteValue).join(",")})`,
-    }
-  }
-  if (!SCALAR_OPERATORS.includes(expr.name)) return null
-  let value = right.value
-  if (expr.name === "like" || expr.name === "ilike") {
-    // TanStack uses SQL %/_ wildcards, but treats * and backslashes literally.
-    // PostgREST always rewrites * to %, so such patterns cannot be pushed.
-    if (typeof value !== "string" || value.includes("*")) return null
-    value = value.replace(/\\/g, "\\\\")
-  }
-  return {
-    column,
-    operator: expr.name,
-    value: quoteScalars ? quoteValue(value) : `${value}`,
-  }
-}
-
-// Embedded expressions are all-or-nothing. Dropping a child of OR or NOT
-// could narrow the response and permanently lose matching rows.
-function toFilterString(
-  expr: Expression,
-  options: FilterOptions,
-  negated = false
-): string | null {
-  if (expr.type !== "func") return null
-  if (expr.name === "not") {
-    return expr.args[0] ? toFilterString(expr.args[0], options, !negated) : null
-  }
-  if (expr.name === "and" || expr.name === "or") {
-    if (expr.args.length === 0) return null
-    // Move NOT down to comparisons with De Morgan's laws. In particular, an
-    // empty IN list needs its null guard even under a negated logical group.
-    const operator = negated ? (expr.name === "and" ? "or" : "and") : expr.name
-    const parts = expr.args.map((arg) => toFilterString(arg, options, negated))
-    return parts.includes(null) ? null : `${operator}(${parts.join(",")})`
-  }
-  const comparison = renderComparison(
-    negated ? { type: "func", name: "not", args: [expr] } : expr,
-    true,
-    options
-  )
-  return comparison
-    ? `${comparison.column}.${comparison.operator}.${comparison.value}`
-    : null
-}
-
-function flattenAnd(expr: Expression): Expression[] {
-  return expr.type === "func" && expr.name === "and"
-    ? expr.args.flatMap(flattenAnd)
-    : [expr]
-}
-
-// Preserve the live adapter's union of duplicate IN demands. This deliberately
-// fetches a superset, which TanStack re-filters. Never do this inside OR/NOT or
-// for server aggregates, where client-side filtering cannot repair the result.
-function mergeInFilters(filters: Expression[]): Expression[] {
-  const merged: Expression[] = []
-  const valuesByField = new Map<string, unknown[]>()
-  for (const filter of filters) {
-    if (filter.type !== "func" || filter.name !== "in") {
-      merged.push(filter)
-      continue
-    }
-    const [left, right] = filter.args
-    if (
-      left?.type !== "ref" ||
-      right?.type !== "val" ||
-      !Array.isArray(right.value)
-    ) {
-      merged.push(filter)
-      continue
-    }
-    const field = JSON.stringify(left.path)
-    const existing = valuesByField.get(field)
-    if (existing) {
-      existing.push(...right.value)
-    } else {
-      const values = [...right.value]
-      valuesByField.set(field, values)
-      merged.push({ ...filter, args: [left, { type: "val", value: values }] })
-    }
-  }
-  return merged
-}
-
-function toPostgrestParams(
-  expr: Expression | undefined | null,
-  options: FilterOptions = {}
-): PostgrestParam[] {
-  if (!expr) return []
-  const conjuncts = flattenAnd(expr)
-  const filters = options.mergeIn ? mergeInFilters(conjuncts) : conjuncts
-  return filters.flatMap((filter): PostgrestParam[] => {
-    const comparison = renderComparison(filter, false, options)
-    if (comparison) return [{ kind: "column", ...comparison }]
-    const embedded = toFilterString(filter, options)
-    if (embedded !== null) {
-      return [
-        {
-          kind: "group",
-          filter: embedded.startsWith("or(") ? embedded.slice(3, -1) : embedded,
-        },
-      ]
-    }
-    if (options.strict) {
-      throw new Error(
-        "Cannot push WHERE expression to PostgREST; server aggregates require fully pushable filters"
-      )
-    }
-    return []
-  })
-}
-
-/** Render params as the `URLSearchParams` sent to (and keyed by) PostgREST. */
-function paramsToSearch(params: PostgrestParam[]): URLSearchParams {
-  const search = new URLSearchParams()
-  for (const param of params) {
-    if (param.kind === "column")
-      search.append(param.column, `${param.operator}.${param.value}`)
-    else search.append("or", `(${param.filter})`)
-  }
-  return search
-}
-
-// Cursor operators arrive pre-flattened as `SimpleComparison`. Scalar values
-// are emitted raw (top-level params must not be quoted); only IN members quote.
-const CURSOR_SCALAR_OPERATORS: Record<string, string> = {
-  eq: "eq",
-  gt: "gt",
-  gte: "gte",
-  lt: "lt",
-  lte: "lte",
-  not_eq: "not.eq",
-}
-
-/** Convert pre-flattened cursor comparisons to PostgREST params. */
-function cursorToPostgrestParams(
-  filters: SimpleComparison[]
-): PostgrestParam[] {
-  return filters.flatMap((filter): PostgrestParam[] => {
-    const column = filter.field.join(".")
-    if (filter.operator === "in") {
-      const values = Array.isArray(filter.value) ? filter.value : []
-      return [
-        {
-          kind: "column",
-          column,
-          operator: "in",
-          value: `(${values.map(quoteValue).join(",")})`,
-        },
-      ]
-    }
-    if (filter.operator === "isNull") {
-      return [{ kind: "column", column, operator: "is", value: "null" }]
-    }
-    const operator = CURSOR_SCALAR_OPERATORS[filter.operator]
-    if (!operator) {
-      console.warn(
-        `cursorToPostgrestParams: unsupported operator: ${filter.operator}`
-      )
-      return []
-    }
-    return [{ kind: "column", column, operator, value: `${filter.value}` }]
-  })
-}
-
-/** Set the comma-joined `order` param (independent of limit/offset). */
-function appendOrder(
-  search: URLSearchParams,
-  sorts: Array<{ column: string; ascending: boolean }>
-): void {
-  if (sorts.length === 0) return
-  search.set(
-    "order",
-    sorts
-      .map((sort) => `${sort.column}.${sort.ascending ? "asc" : "desc"}`)
-      .join(",")
-  )
-}
-
-function appendLimit(search: URLSearchParams, limit: number): void {
-  search.set("limit", `${limit}`)
-}
-
-function appendOffset(search: URLSearchParams, offset: number): void {
-  search.set("offset", `${offset}`)
-}
-
-/**
- * Encode the params that identify a cached subset — where filters, order, and
- * limit. Deliberately excludes `select`, `cursor`, and `offset`: those are the
- * per-page request details, not part of what distinguishes one subset from
- * another. This is the shared core of the request URL and the query key, so the
- * two cannot drift.
- */
-export function subsetParamsToSearch(
-  options: LoadSubsetOptions
-): URLSearchParams {
-  const { where, orderBy, limit } = options
-  const search = paramsToSearch(toPostgrestParams(where, { mergeIn: true }))
-  appendOrder(
-    search,
-    parseOrderByExpression(orderBy).map((sort) => ({
-      column: sort.field.join("."),
-      ascending: sort.direction === "asc",
-    }))
-  )
-  if (limit) {
-    appendLimit(search, limit)
-  }
-  return search
-}
-
-/**
- * Build the full read query string for a TanStack DB subset load: the shared
- * subset params ({@link subsetParamsToSearch}) plus `select=*`, cursor filters,
- * and offset.
- */
-export function loadSubsetOptionsToSearch(
-  options: LoadSubsetOptions
-): URLSearchParams {
-  const search = new URLSearchParams()
-  search.set("select", "*")
-  for (const [key, value] of subsetParamsToSearch(options)) {
-    search.append(key, value)
-  }
-
-  const cursorFilters = options.cursor
-    ? [...extractSimpleComparisons(options.cursor.whereFrom)]
-    : []
-  for (const [key, value] of paramsToSearch(
-    cursorToPostgrestParams(cursorFilters)
-  )) {
-    search.append(key, value)
-  }
-
-  if (options.offset) {
-    appendOffset(search, options.offset)
-  }
-  return search
-}
-
-/** Build the `key.eq.value` search that matches one row by its key columns. */
-export function keyColumnsToSearch(
-  keys: string[],
-  item: Record<string, unknown>
-): URLSearchParams {
-  return paramsToSearch(
-    keys.map((key) => ({
-      kind: "column",
-      column: key,
-      operator: "eq",
-      value: `${item[key]}`,
-    }))
-  )
-}
+} from "../serialize"
+import {
+  appendLimit,
+  appendOffset,
+  appendOrder,
+  paramsToSearch,
+  toPostgrestParams,
+} from "./common"
 
 // ── Serialized query IR → search ────────────────────────────────────
